move RPC data structures from core to vcell-api-types
