+++ vcell-api-types/src/main/java/org/vcell/api/types/rpc/VCellApiRpcBody.java
@@ -1,4 +1,4 @@
-package cbit.rmi.event;
+package org.vcell.api.types.rpc;
 
 import java.io.Serializable;
