+++ vcell-apiclient/src/main/java/org/vcell/api/messaging/RemoteProxyVCellConnectionFactory.java
@@ -26,8 +26,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.vcell.DependencyConstants;
 import org.vcell.api.client.VCellApiClient;
-import cbit.rmi.event.VCellApiRpcBody;
-import cbit.rmi.event.VCellApiRpcRequest;
+import org.vcell.api.types.rpc.VCellApiRpcBody;
+import org.vcell.api.types.rpc.VCellApiRpcRequest;
 import org.vcell.api.utils.Auth0ConnectionUtils;
 import org.vcell.restclient.ApiException;
 import org.vcell.restclient.model.AccesTokenRepresentationRecord;
